+++ cmd/neofs-cli/modules/control/object.go
@@ -1,0 +1,18 @@
+package control
+
+import (
+	"github.com/spf13/cobra"
+)
+
+var objectCmd = &cobra.Command{
+	Use:   "object",
+	Short: "Direct object operations with storage engine",
+}
+
+func initControlObjectsCmd() {
+	objectCmd.AddCommand(listObjectsCmd)
+	objectCmd.AddCommand(objectStatusCmd)
+
+	initControlObjectsListCmd()
+	initObjectStatusFlags()
+}